LayoutTests/http/tests/preload/css-import-base-tag-expected.txt

Test that the preload scanner correctly resolves relative @import URLs in CSS against a base tag

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isPreloaded('http://127.0.0.1:8000/preload/resources/base-tag-import.css'); is true
PASS successfullyParsed is true

TEST COMPLETE
